Adds backup and restore data functionality
Implements data backup and restore functionality to allow users to save and restore their drawing projects.

Adds options to the profile settings to export all Hive data to a JSON file in the downloads directory and import data from a selected JSON file, facilitating data migration and recovery.

The update also improves the watch view by adding a refresh indicator to the comment list and displays the comment creation date and time.