Enables data restoration and project loading
Facilitates the restoration of Hive data and subsequent loading of projects by integrating the home controller within the profile controller.
This ensures that the UI is updated immediately after a data restoration, providing a seamless user experience.
Also, it initializes the Home Controller as a singleton service for global access.